Use component props
