+++ examples/todoapp/src/index.css
@@ -1,3 +1,8 @@
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
 }
+
+.fabrix.table {
+  max-width: 75em;  
+  margin: auto;
+}
